+++ src/compile.h
@@ -8,8 +8,12 @@
 
 #include "utils.h"
 
+// orginal clang compiler main entry
+int iclang_main(int argc, const char **argv);
+
 namespace icpp {
 
+// orignal clang-format main entry
 int cformat_main(int argc, const char **argv);
 
 int compile_source_clang(int argc, const char **argv, bool cl = false);